requirements.txt
# ── Core ──────────────────────────────────────────────
fastapi==0.110.0
uvicorn[standard]==0.29.0
python-dotenv==1.0.1
pydantic==2.6.4

# ── Data & ML ─────────────────────────────────────────
pandas==2.2.1
numpy==1.26.4
scikit-learn==1.4.1
scipy==1.13.0
scikit-surprise==1.1.3   # collaborative filtering (SVD)

# ── Database ──────────────────────────────────────────
sqlalchemy==2.0.29
aiosqlite==0.20.0          # async SQLite
alembic==1.13.1            # migrations

# ── HTTP / API ────────────────────────────────────────
httpx==0.27.0              # async TMDB calls
requests==2.31.0

# ── Utilities ─────────────────────────────────────────
joblib==1.3.2              # model serialization
python-multipart==0.0.9
slowapi==0.1.9             # rate limiting

# ── Optional Deep Learning ────────────────────────────
# tensorflow==2.16.1
# torch==2.2.2
